fix: move functions to module that uses them
